refactor(calendar): extract shared view primitives
