+++ src/components/calendar-keyboard.test.ts
@@ -2,10 +2,10 @@
 import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
 
-const source = readFileSync(
-  new URL("./calendar-view.tsx", import.meta.url),
-  "utf8",
-);
+const source = [
+  readFileSync(new URL("./calendar-view.tsx", import.meta.url), "utf8"),
+  readFileSync(new URL("./calendar-view-primitives.tsx", import.meta.url), "utf8"),
+].join("\n");
 
 assert.match(
   source,
